fix(#1509): break gate_record import cycle (architecture no-new-cycles)
Replace `from scistudio.qa.governance.gate_record import <sibling>` with
direct `import scistudio.qa.governance.gate_record.<sibling> as <sibling>`
in cli.py, evaluator.py, workflow.py, checks.py, and all guards/*.py so
submodules never import through the package __init__, eliminating the 16-module
SCC. Extract GuardInputs/Guard into guards/_base.py (leaf module) so guard
submodules import GuardInputs from _base directly instead of from guards/__init__,
eliminating the 11-module guards SCC. Cycle count: 5 -> 4 (back to baseline).

Refs #1509
