+++ src/scistudio/qa/governance/gate_record/guards/__init__.py
@@ -15,76 +15,22 @@ interface signature is frozen so B2/B3/B4 build against it without churn.
 
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping, Sequence
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any
+# GuardInputs and Guard live in _base to avoid import cycles: guard submodules
+# import from _base directly; __init__ re-exports them for external callers.
+import scistudio.qa.governance.gate_record.guards.core_change_guard as core_change_guard
+import scistudio.qa.governance.gate_record.guards.docs_landing as docs_landing
+import scistudio.qa.governance.gate_record.guards.human_bypass_guard as human_bypass_guard
+import scistudio.qa.governance.gate_record.guards.issue_link as issue_link
+import scistudio.qa.governance.gate_record.guards.mod_guard as mod_guard
+import scistudio.qa.governance.gate_record.guards.persona_policy as persona_policy
+import scistudio.qa.governance.gate_record.guards.pr_merge_guard as pr_merge_guard
+import scistudio.qa.governance.gate_record.guards.sentrux_gate as sentrux_gate
+import scistudio.qa.governance.gate_record.guards.test_engineer_scope_guard as test_engineer_scope_guard
+import scistudio.qa.governance.gate_record.guards.weakened_ci_check as weakened_ci_check
 
-from scistudio.qa.governance.gate_record.ledger import (
-    AdminLabel,
-    IssueRef,
-    Persona,
-    StrictnessTier,
-    TaskKind,
-)
-from scistudio.qa.schemas.report import AuditReport
-
-
-@dataclass(frozen=True)
-class GuardInputs:
-    """The evaluator-built input bundle every guard calculator receives (§4).
-
-    Guards must read ONLY from this bundle. They must not load the ledger, run
-    git, or maintain independent rule sets.
-    """
-
-    repo_root: Path
-    mode: str
-    task_kind: TaskKind
-    persona: Persona
-    runtime: str
-    tier: StrictnessTier
-    governance_touch: bool
-    # Observed (git-derived) facts.
-    changed_files: Sequence[str]
-    diff_fingerprint: str | None
-    # Classified surfaces: surface-class name -> changed paths in that class.
-    surfaces: Mapping[str, Sequence[str]]
-    # Effective declared scope.
-    effective_include: Sequence[str]
-    effective_exclude: Sequence[str]
-    # Declared vs observed docs/tests (already reconciled by the evaluator).
-    declared_docs_paths: Sequence[str]
-    declared_test_paths: Sequence[str]
-    verified_docs_paths: Sequence[str]
-    verified_test_paths: Sequence[str]
-    # Issues and PR facts.
-    issues: Sequence[IssueRef]
-    pr_body: str | None
-    pr_context: Mapping[str, Any] | None
-    # Admin labels.
-    requested_admin_labels: Sequence[AdminLabel]
-    observed_admin_labels: Sequence[AdminLabel]
-    # Free-form extras the evaluator may attach (e.g. sentrux evidence).
-    extras: Mapping[str, Any] = field(default_factory=dict)
-
-
-# A guard calculator: pure function from inputs to an AuditReport.
-Guard = Callable[[GuardInputs], AuditReport]
-
-
-from scistudio.qa.governance.gate_record.guards import (  # noqa: E402  (avoid cycle at import top)
-    core_change_guard,
-    docs_landing,
-    human_bypass_guard,
-    issue_link,
-    mod_guard,
-    persona_policy,
-    pr_merge_guard,
-    sentrux_gate,
-    test_engineer_scope_guard,
-    weakened_ci_check,
-)
+# GuardInputs and Guard live in _base to avoid import cycles: guard submodules
+# import from _base directly; __init__ re-exports them for external callers.
+from scistudio.qa.governance.gate_record.guards._base import Guard, GuardInputs
 
 # The registry the evaluator iterates. Each guard runs exactly once. B2 (#1509)
 # replaces these stubs with real calculators; the keys and signatures are
