+++ src/scistudio/qa/governance/gate_record/guards/core_change_guard.py
@@ -15,8 +15,8 @@ from __future__ import annotations
 from collections.abc import Mapping, Sequence
 from typing import Any
 
-from scistudio.qa.governance.gate_record import labels as label_vocab
-from scistudio.qa.governance.gate_record.guards import GuardInputs
+import scistudio.qa.governance.gate_record.labels as label_vocab
+from scistudio.qa.governance.gate_record.guards._base import GuardInputs
 from scistudio.qa.governance.gate_record.guards._stub import source_sha
 from scistudio.qa.schemas.report import AuditReport, AuditStatus, Finding, Severity
