+++ src/scistudio/qa/governance/gate_record/guards/issue_link.py
@@ -15,8 +15,8 @@ from __future__ import annotations
 
 import re
 
-from scistudio.qa.governance.gate_record import surfaces
-from scistudio.qa.governance.gate_record.guards import GuardInputs
+import scistudio.qa.governance.gate_record.surfaces as surfaces
+from scistudio.qa.governance.gate_record.guards._base import GuardInputs
 from scistudio.qa.governance.gate_record.guards._stub import source_sha
 from scistudio.qa.schemas.report import AuditReport, AuditStatus, Finding, Severity
