+++ src/scistudio/qa/governance/gate_record/guards/mod_guard.py
@@ -15,8 +15,8 @@ through the ledger ``governance_touch`` flag plus ``requested`` (local intent) /
 
 from __future__ import annotations
 
-from scistudio.qa.governance.gate_record import labels as label_vocab
-from scistudio.qa.governance.gate_record.guards import GuardInputs
+import scistudio.qa.governance.gate_record.labels as label_vocab
+from scistudio.qa.governance.gate_record.guards._base import GuardInputs
 from scistudio.qa.governance.gate_record.guards._stub import source_sha
 from scistudio.qa.schemas.report import AuditReport, AuditStatus, Finding, Severity
