+++ src/scistudio/qa/governance/gate_record/guards/pr_merge_guard.py
@@ -12,8 +12,8 @@ itself. The guard only fires in CI mode, where a real merge event exists.
 
 from __future__ import annotations
 
-from scistudio.qa.governance.gate_record import labels as label_vocab
-from scistudio.qa.governance.gate_record.guards import GuardInputs
+import scistudio.qa.governance.gate_record.labels as label_vocab
+from scistudio.qa.governance.gate_record.guards._base import GuardInputs
 from scistudio.qa.governance.gate_record.guards._stub import source_sha
 from scistudio.qa.schemas.report import AuditReport, AuditStatus, Finding, Severity
