+++ src/scistudio/qa/governance/gate_record/guards/sentrux_gate.py
@@ -24,8 +24,8 @@ from typing import Any, cast
 
 from pydantic import AliasChoices, BaseModel, ConfigDict, Field, ValidationError, field_validator
 
-from scistudio.qa.governance.gate_record import surfaces
-from scistudio.qa.governance.gate_record.guards import GuardInputs
+import scistudio.qa.governance.gate_record.surfaces as surfaces
+from scistudio.qa.governance.gate_record.guards._base import GuardInputs
 from scistudio.qa.governance.gate_record.guards._stub import source_sha
 from scistudio.qa.schemas.report import AuditReport, AuditStatus, Finding, Severity
